Add files via upload
Update architecture of the project add some general function in com/example/utils/Utils and cleaned other objects
